+++ vcf/src/vcf/FoundationKit/BasicOutputStream.h
@@ -42,7 +42,7 @@ public:
 
 	virtual uint64 getCurrentSeekPos();
 
-	virtual uint32 write( const unsigned char* bytesToRead, uint32 sizeOfBytes );
+	virtual uint64 write( const unsigned char* bytesToRead, uint64 sizeOfBytes );
 
 	void init();
 private:
